+++ src/ztc/datamodel/admin/resultaattype.py
@@ -2,7 +2,8 @@ from django.contrib import admin
 from django.utils.translation import ugettext_lazy as _
 
 from ..models import ResultaatType, ZaakInformatieobjectTypeArchiefregime
-from .mixins import FilterSearchOrderingAdminMixin, GeldigheidAdminMixin
+from .forms import ResultaatTypeForm
+from .mixins import GeldigheidAdminMixin
 
 
 class ZaakInformatieobjectTypeArchiefregimeInline(admin.TabularInline):
@@ -11,37 +12,51 @@ class ZaakInformatieobjectTypeArchiefregimeInline(admin.TabularInline):
 
 
 @admin.register(ResultaatType)
-class ResultaatTypeAdmin(GeldigheidAdminMixin, FilterSearchOrderingAdminMixin, admin.ModelAdmin):
+class ResultaatTypeAdmin(GeldigheidAdminMixin, admin.ModelAdmin):
     model = ResultaatType
+    form = ResultaatTypeForm
 
     # List
-    list_display = ('resultaattypeomschrijving', 'selectielijstklasse')
+    list_display = ('omschrijving', 'omschrijving_generiek', 'selectielijstklasse', 'uuid')
+    ordering = ('zaaktype', 'omschrijving')
+    search_fields = (
+        'omschrijving',
+        'omschrijving_generiek',
+        'selectielijstklasse',
+        'toelichting',
+        'uuid',
+    )
 
     # Details
     fieldsets = (
         (_('Algemeen'), {
             'fields': (
-                'resultaattypeomschrijving',
-                'resultaattypeomschrijving_generiek',
-                'selectielijstklasse',
-                'archiefnominatie',
-                'archiefactietermijn',
-                'brondatum_archiefprocedure',
+                'zaaktype',
+                'omschrijving',
                 'toelichting',
-                'heeft_voor_brondatum_archiefprocedure_relevante',
-                'is_relevant_voor'
             )
         }),
-        (_('Relaties'), {
+        (_('Gemeentelijke selectielijst'), {
             'fields': (
-                'heeft_verplichte_zot',
-                'heeft_verplichte_ziot'
+                'resultaattypeomschrijving',
+                'selectielijstklasse',
             )
         }),
+        (_('Bepaling brondatum archiefprocedure'), {
+            'fields': (
+                'brondatum_archiefprocedure_afleidingswijze',
+                'brondatum_archiefprocedure_datumkenmerk',
+                'brondatum_archiefprocedure_einddatum_bekend',
+                'brondatum_archiefprocedure_objecttype',
+                'brondatum_archiefprocedure_registratie',
+                'brondatum_archiefprocedure_procestermijn',
+            ),
+        }),
+        # (_('Relaties'), {
+        #     'fields': (
+        #         'heeft_verplichte_zot',
+        #         'heeft_verplichte_ziot'
+        #     )
+        # }),
     )
-    filter_horizontal = (
-        'heeft_verplichte_zot',
-        'heeft_verplichte_ziot',
-    )
-    raw_id_fields = ('heeft_voor_brondatum_archiefprocedure_relevante', 'is_relevant_voor', )
-    inlines = (ZaakInformatieobjectTypeArchiefregimeInline,)  # 'bepaalt_afwijkend_archiefregime_van',
+    raw_id_fields = ('zaaktype',)